Get og Set kan sende 16 bits integers

+++ P3-325c70fd5aebbd129a3583fdb97339f83d80e37f/Serial/Serial-PC/enc_temp_folder/baa38aa7ffd4d08be95c83a31e4dd7a7/Control.h
@@ -8,8 +8,8 @@ public:
     Serial* SP = new Serial("\\\\.\\COM12");    // adjust as needed
     double getTorque() {};
     void setTorque(double goalTorque, char Data[256], int motorID); // Måske ikke lav
-    int16_t getPosition(int motorID, char Data[256]); // Position in radians
-    void setPosition(int16_t goalPos, char Data[256], int motorID); // Måske ikke lav
+    double getPosition(int motorID, char Data[256]); // Position in radians
+    void setPosition(double goalPos, char Data[256], int motorID); // Måske ikke lav
     double getVelocity(int motorID, char Data[256]);
     void setVelocity(double goalVel, char Data[256], int motorID);
     double getAcceleration() {};
